Added class to manage game variables
All game variables are stored in sessions which are updated as the game plays. When resetting the game either from logging out or starting a new game, each session attribute had to be called individually and updated. Now the game_variables class contains all game essential variables and the init method to reset them all to the starting position. This is then lopped through with the reset_game_attributes method and assigned to sessions accordingly.