Update riddles.json and allow for multiple answers
Previously the riddles.json file contained only dummy text, this
has been updated with actual riddles and the relevant answers.

An issue became apparent that concerned checking whether a user's
input was correct and corresponded with the riddles.json file. For
example, if the answer was 'television' and the user inputted 'a
television', this would be incorrectly flagged as wrong.

To combat this, a list of answers are supplied with each riddle,
covering the most popular variations users may use to answer
correctly. These are looped through to ensure the user has
entered one variation of a correct answer.

The check to ensure a user's answer is correct and which message
to show now takes place in the run.py (where previously it was
completed in the questions.html file).