Add Python testing document to test main functionality
The testing document uses unittest to ensure the functionality of the game works as intended.
Tests include ensuring all pages return a 200 response when the appropriate attributes are
met, the total time it takes a user to answer a question is accurately recorded, tests to
ensure users cannot skip ahead in the game before completing necessary steps or questions,
and ensuring visited-check is updated as intended.